Struct amethyst::asset_manager::AssetManager

pub struct AssetManager { /* fields omitted */ }

Asset manager which handles assets and loaders.

Methods

impl AssetManager

fn new() -> AssetManager

Create a new asset manager

fn register_loader<A: Any + Send + Sync, S: Any>(&mut self, asset: &str) where S: AssetLoader<A> + AssetLoaderRaw

Register a new loading method for a specific asset data type

fn register_store<T: 'static + AssetStore>(&mut self, store: T)

Register an asset store

fn load_asset_from_raw<A: Any + Send + Sync>(&mut self,
name: &str,
asset_type: &str,
raw: &[u8])
-> Option<AssetId>

Load an asset from raw data

Panics

Panics if the asset type isn't registered

fn load_asset<A: Any + Send + Sync>(&mut self,
name: &str,
asset_type: &str)
-> Option<AssetId>

Load an asset from the asset stores

fn create_renderable(&self,
mesh: &str,
ka: &str,
kd: &str,
ks: &str,
ns: f32)
-> Option<Renderable>

Create a Renderable component from a loaded mesh and ka/kd/ks textures