Module amethyst::assets::formats::textures

Formats to load textures Provides texture formats

Structs

BmpFormat
Allows loading of BMP files.

ImageData
ImageData provided by formats, can be interpreted as a texture.

JpegFormat
Allows loading of Jpeg files.

JpgFormat
Allows loading of Jpg files.

PngFormat
Allows loading of PNG files.

Enums

ImageError
Error.

Type Definitions

ImageFuture
A future which will eventually have an image available.