Struct amethyst::audio::Dj

pub struct Dj { /* fields omitted */ }

This structure provides a way to programmatically pick and play music.

Methods

impl Dj

pub fn new(output: &Output) -> Dj

Creates a new Dj using the given audio output.

pub fn set_picker(&mut self, picker: Box<FnMut(&mut Dj) -> bool + Send + Sync>)

A Dj's picker will be called by the DjSystem whenever the Dj runs out of music to play.

Only the Dj added to the world's resources with resource ID 0 will have their picker called.

During callback the picker is separated from the Dj in order to avoid multiple aliasing. After the callback is complete, if the picker returned true it will be reattached.

pub fn clear_picker(&mut self)

Clears the previously set picker.

pub fn append(&self, source: &Source) -> Result<(), DecoderError>

Adds a source to the Dj's queue of music to play.

pub fn empty(&self) -> bool

Returns true if the Dj has no more music to play.

pub fn volume(&self) -> f32

Retrieves the volume of the Dj, between 0.0 and 1.0;

pub fn set_volume(&mut self, volume: f32)

Sets the volume of the Dj.

pub fn play(&self)

Resumes playback of a paused Dj. Has no effect if this Dj was never paused.

pub fn pause(&self)

Pauses playback, this can be resumed with Dj::play

pub fn is_paused(&self) -> bool

Returns true if the Dj is currently paused.

pub fn stop(&self)

Empties the Dj's queue of all music.