Struct amethyst::ecs::changeset::ChangeSet

pub struct ChangeSet<T> { /* fields omitted */ }

Change set that can be collected from an iterator, and joined on for easy application to components.

Example

pub struct Health(i32);

impl Component for Health {
    type Storage = DenseVecStorage<Self>;
}

let a = world.create_entity().with(Health(100)).build();
let b = world.create_entity().with(Health(200)).build();

let changeset = [(a, 32), (b, 12), (b, 13)]
    .iter()
    .cloned()
    .collect::<ChangeSet<i32>>();
for (health, modifier) in (&mut world.write_storage::<Health>(), &changeset).join() {
    health.0 -= modifier;
}

Methods

impl<T> ChangeSet<T>

pub fn new() -> ChangeSet<T>

Create a new change set

pub fn add(&mut self, entity: Entity, value: T) where
T: AddAssign<T>,

Add a value to the change set. If the entity already have a value in the change set, the incoming value will be added to that.

pub fn clear(&mut self)

Clear the changeset