Struct amethyst::ecs::components::Renderable

pub struct Renderable {
    pub mesh: Mesh,
    pub ambient: Texture,
    pub diffuse: Texture,
    pub specular: Texture,
    pub specular_exponent: f32,
}

A Component that can be attached to an ECS Entity to render it onscreen.

It combines geometry and various textures used in lighting calculations to represent an entity on the screen. Every Renderable, Transform pair attached to an entity inside the ECS World is rendered by the GfxDevice::render_world method.

Fields
mesh: Mesh

Mesh asset to be renderered.

ambient: Texture

Applied during ambient lighting calculations.

diffuse: Texture

Applied during diffuse lighting calculations.

specular: Texture

Applied during specular lighting calculations.

specular_exponent: f32

Shininess of the object's surface.

Methods

impl Renderable

fn new(mesh: Mesh,
ambient: Texture,
diffuse: Texture,
specular: Texture,
specular_exponent: f32)
-> Renderable

Creates a new renderable. You will probably want not use this directly. Instead, use the AssetManager::create_renderable function.