Struct amethyst::ecs::input::bundle::InputBundle

pub struct InputBundle<AX, AC> where
    AX: Hash + Eq,
    AC: Hash + Eq,  { /* fields omitted */ }

Bundle for adding the InputHandler.

This also adds the Winit EventHandler and the InputEvent EventHandler where AC is the type for Actions you have assigned here.

Type parameters

AX: The type used to identify input axes. AC: The type used to identify input actions.

String is appropriate for either of these if you don't know what to use.

Errors

No errors returned from this bundle.

Methods

impl<AX, AC> InputBundle<AX, AC> where
AX: Hash + Eq + DeserializeOwned + Serialize + Default,
AC: Hash + Eq + DeserializeOwned + Serialize + Default,

fn new() -> Self

Create a new input bundle with no bindings

fn with_bindings(self, bindings: Bindings<AX, AC>) -> Self

Use the provided bindings with the InputHandler

fn with_bindings_from_file<P: AsRef<Path>>(self, file: P) -> Self

Load bindings from file