Struct amethyst::ecs::input::Bindings

pub struct Bindings<AX, AC> where
    AC: Hash + Eq,
    AX: Hash + Eq,  { /* fields omitted */ }

Used for saving and loading input settings.

Methods

impl<AX, AC> Bindings<AX, AC> where
AC: Hash + Eq + Clone,
AX: Hash + Eq + Clone,

fn new() -> Bindings<AX, AC>

Creates a new empty Bindings structure

fn insert_axis<A>(&mut self, id: A, axis: Axis) -> Option<Axis> where
A: Into<AX>,

Assign an axis to an ID value

This will insert a new axis if no entry for this id exists. If one does exist this will replace the axis at that id and return it.

fn remove_axis<A>(&mut self, id: &A) -> Option<Axis> where
A: Eq + Hash + ?Sized,
AX: Borrow<A>,

Removes an axis, this will return the removed axis if successful.

fn axis<A>(&mut self, id: &A) -> Option<&Axis> where
A: Eq + Hash + ?Sized,
AX: Borrow<A>,

Returns a reference to an axis.

fn axes(&self) -> Vec<AX>

Gets a list of all axes

fn insert_action_binding<A>(&mut self, id: A, binding: Button) where
A: Hash + Eq + Into<AC>,
AC: Borrow<A>,

Add a button to an action.

This will insert a new binding between this action and the button.

fn remove_action_binding<T>(&mut self, id: &T, binding: Button) where
AC: Borrow<T>,
T: Eq + Hash + ?Sized,

Removes an action binding that was assigned previously.

fn action_bindings<T>(&self, id: &T) -> Option<&[Button]> where
AC: Borrow<T>,
T: Eq + Hash + ?Sized,

Returns an action's bindings.

fn actions(&self) -> Vec<AC>

Gets a list of all action bindings