Struct amethyst::ecs::rendering::bundle::RenderBundle

pub struct RenderBundle<P> { /* fields omitted */ }

Rendering bundle

Will register all necessary components needed for rendering, along with any resources. Will also register asset contexts with the asset Loader, and add systems for merging AssetFuture into its related component.

Will add RenderSystem as a thread local system.

Errors

Returns errors related to:

Renderer creation

Pipeline creation

Methods

impl<P> RenderBundle<P> where
P: PipelineBuild,

fn new(pipe: P) -> Self

Create a new render bundle with the given pipeline

fn with_config(self, config: DisplayConfig) -> Self

Use the given display config for configuring window and render properties