Module amethyst::ecs::rendering::resources

amethyst rendering ecs resources

Structs

AmbientColor
The ambient color of a scene

Factory
The factory abstraction, which allows to access the real factory and returns futures.

FactoryFuture
A factory future.

Type Definitions

MaterialFuture
A material which may not have been created yet.

MeshFuture
A mesh which may not have been created yet.

TextureFuture
A texture which may not have been created yet.