Struct amethyst::ecs::rendering::resources::Factory

pub struct Factory { /* fields omitted */ }

The factory abstraction, which allows to access the real factory and returns futures.

Methods

impl Factory

pub fn new() -> Self

Creates a new factory resource.

pub fn create_mesh<D, V>(&self, mb: MeshBuilder<D, V>) -> MeshFuture where
D: AsRef<[V]> + Send + Sync + 'static,
V: VertexFormat + Send + Sync + 'static,

Creates a mesh asynchronously.

pub fn create_texture<D, T>(&self, tb: TextureBuilder<D, T>) -> TextureFuture where
D: AsRef<[T]> + Send + Sync + 'static,
T: Pod + Send + Sync + 'static,

Creates a texture asynchronously.

pub fn create_material<DA, TA, DE, TE, DN, TN, DM, TM, DR, TR, DO, TO, DC, TC>(
&self,
mb: MaterialBuilder<DA, TA, DE, TE, DN, TN, DM, TM, DR, TR, DO, TO, DC, TC>
) -> MaterialFuture where
DA: AsRef<[TA]> + Send + Sync + 'static,
TA: Pod + Send + Sync + 'static,
DE: AsRef<[TE]> + Send + Sync + 'static,
TE: Pod + Send + Sync + 'static,
DN: AsRef<[TN]> + Send + Sync + 'static,
TN: Pod + Send + Sync + 'static,
DM: AsRef<[TM]> + Send + Sync + 'static,
TM: Pod + Send + Sync + 'static,
DR: AsRef<[TR]> + Send + Sync + 'static,
TR: Pod + Send + Sync + 'static,
DO: AsRef<[TO]> + Send + Sync + 'static,
TO: Pod + Send + Sync + 'static,
DC: AsRef<[TC]> + Send + Sync + 'static,
TC: Pod + Send + Sync + 'static,

Creates a mesh asynchronously.

pub fn execute<F, T, E>(&self, fun: F) -> FactoryFuture<T, E> where
F: FnOnce(&mut Factory) -> Result<T, E> + Send + Sync + 'static,
T: Send + Sync + 'static,
E: Send + Sync + 'static,

Execute a closure which takes in the real factory.