Module amethyst::ecs::resources

Resources that can be added to ecs::World.

Camera, ScreenDimensions, and Time are added by default and automatically updated every frame by Application.

Structs

Broadcaster
Allows publishing entities

Camera
Represents a camera looking around inside a game world.

InputHandler
This struct holds state information about input devices.

ScreenDimensions
Abstract representation of a screen.

Time
Time is added to ecs::World as a resource by default. It is updated every frame in Application::advance_frame.

Enums

Projection
Represents the graphical projection of a Camera.