Struct amethyst::ecs::resources::InputHandler

pub struct InputHandler { /* fields omitted */ }

This struct holds state information about input devices.

For example, if a key is pressed on the keyboard, this struct will record that the key is pressed until it is released again. Usage requires pumping input events through this struct, using the following code in the handle_events method of State.

fn handle_events(&mut self,
                 events: &[WindowEvent],
                 _: &mut World,
                 _: &mut AssetManager,
                 _: &mut Pipeline)
                 -> Trans {
    // ...
    let mut input_handler = world.write_resource::<InputHandler>();
    input_handler.update(events);
    // ...
}

Methods

impl InputHandler

fn new() -> InputHandler

Creates a new input handler.

fn update(&mut self, events: &[WindowEvent])

Updates the input handler with new engine events.

fn text_entered(&self) -> &str

Returns a string representation of all text entered this frame.

Intended for use with text entry fields, insert this string at the cursor position every frame.

fn pressed_keys(&self) -> PressedKeys

Returns an iterator for all the pressed down keys.

fn key_is_pressed(&self, key: VirtualKeyCode) -> bool

Checks if the given key is being pressed.

fn keys_are_pressed(&self, keys: &[VirtualKeyCode]) -> bool

Checks if all the given keys are being pressed.

fn key_down(&self, key: VirtualKeyCode) -> bool

Checks if the given key was pressed on this frame.

fn key_released(&self, key: VirtualKeyCode) -> bool

Checks if a given key was released on this frame.

fn keys_down(&self, keys: &[VirtualKeyCode]) -> bool

Checks if the all the given keys are down and at least one was pressed on this frame.

fn pressed_mouse_buttons(&self) -> PressedMouseButtons

Returns an iterator for all the pressed down mouse buttons.

fn mouse_button_is_pressed(&self, button: MouseButton) -> bool

Checks if the given mouse button is being pressed.

fn mouse_buttons_are_pressed(&self, buttons: &[MouseButton]) -> bool

Checks if all the given mouse buttons are being pressed.

fn mouse_button_down(&self, button: MouseButton) -> bool

Checks if the given mouse button was pressed this frame.

fn mouse_button_released(&self, button: MouseButton) -> bool

Checks if the given mouse button was released this frame.

fn mouse_buttons_down(&self, buttons: &[MouseButton]) -> bool

Checks if the all the given mouse buttons are down and at least one was pressed this frame.

fn mouse_position(&self) -> Option<(i32, i32)>

Gets the current mouse position.

this method can return None, either if no mouse is connected, or if no mouse events have been recorded

fn mouse_position_change(&self) -> (i32, i32)

Gets the change in position since the last frame.

fn pressed_buttons(&self) -> PressedButtons

Returns an iterator over the buttons that are currently pressed

fn button_is_pressed(&self, button: Button) -> bool

Checks if the given button is currently pressed.

fn buttons_are_pressed(&self, buttons: &[Button]) -> bool

Checks if all the given buttons are pressed.

fn button_down(&self, button: Button) -> bool

Checks if the given button was pressed on this frame.

fn button_released(&self, button: Button) -> bool

Checks if the given button was released on this frame.

fn buttons_down(&self, buttons: &[Button]) -> bool

Checks if the all the given buttons are being pressed and at least one was pressed this frame.

fn axis_value(&self, id: i32) -> Option<f32>

Returns the value of an axis by the i32 id, if the id doesn't exist this returns None.

fn action_is_pressed(&self, action: i32) -> Option<bool>

Checks if the given button is currently pressed.

fn actions_are_pressed(&self, actions: &[i32]) -> Result<bool, Vec<i32>>

Checks if all the given actions are pressed.

If any action in this list is invalid this will return the id of it in Err.

fn action_down(&self, action: i32) -> Option<bool>

Checks if the given action was pressed on this frame.

fn action_released(&self, action: i32) -> Option<bool>

Checks if the given action was released on this frame.

fn actions_down(&self, actions: &[i32]) -> Result<bool, Vec<i32>>

Checks if the all the given actions are being pressed and at least one was pressed this frame.

If any action in this list is invalid this will return the id of it in Err.

fn insert_axis(&mut self, id: i32, axis: Axis) -> Option<Axis>

Assign an axis to an ID value

This will insert a new axis if no entry for this id exists. If one does exist this will replace the axis at that id and return it.

fn remove_axis(&mut self, id: i32) -> Option<Axis>

Removes an axis, this will return the removed axis if successful.

fn get_axis(&mut self, id: i32) -> Option<&Axis>

Returns a reference to an axis.

fn insert_action(&mut self, id: i32, binding: Button) -> Option<Button>

Assign an action to an ID value

This will insert a new action if no entry for this id exists. If one does exist this will replace the action at that id and return it.

fn remove_action(&mut self, id: i32) -> Option<Button>

Removes an action, this will return the removed action if successful.

fn get_action(&mut self, id: i32) -> Option<&Button>

Returns a reference to an action's button.