Struct amethyst::ecs::resources::ScreenDimensions

pub struct ScreenDimensions {
    pub w: f32,
    pub h: f32,
    pub aspect_ratio: f32,
}

Abstract representation of a screen.

Fields
w: f32

Screen width in pixels (px).

h: f32

Screen height in pixels (px).

aspect_ratio: f32

Width divided by height.

Methods

impl ScreenDimensions

fn new(w: u32, h: u32) -> ScreenDimensions

Creates a new screen dimensions object with the given width and height.

fn update(&mut self, w: u32, h: u32)

Updates the width and height of the screen and recomputes the aspect ratio.