Struct amethyst::ecs::storage::RestrictedStorage

pub struct RestrictedStorage<'rf, 'st, C, S, B, Restrict> where
    'st: 'rf,
    B: Borrow<BitSet> + 'rf,
    C: Component,
    S: Borrow<<C as Component>::Storage> + 'rf,  { /* fields omitted */ }

Similar to a MaskedStorage and a Storage combined, but restricts usage to only getting and modifying the components. That means nothing that would modify the inner bitset so the iteration cannot be invalidated. For example, no insertion or removal is allowed.

Example Usage:

struct SomeComp(u32);
impl Component for SomeComp {
    type Storage = VecStorage<Self>;
}

struct RestrictedSystem;
impl<'a> System<'a> for RestrictedSystem {
    type SystemData = (
        Entities<'a>,
        WriteStorage<'a, SomeComp>,
    );
    fn run(&mut self, (entities, mut some_comps): Self::SystemData) {
        for (entity, mut comps) in (
            &entities,
            &mut some_comps.restrict_mut()
        ).join() {
            // Check if the reference is fine to mutate.
            if comps.get_unchecked().0 < 5 {
                // Get a mutable reference now.
                let mut mutable = comps.get_mut_unchecked();
                mutable.0 += 1;
            }
        }
    }
}