Struct amethyst::ecs::transform::components::LocalTransform

pub struct LocalTransform { /* fields omitted */ }

Local position, rotation, and scale (from parent if it exists).

Used for rendering position and orientation.

Methods

impl LocalTransform

pub fn flag(&self, dirty: bool)

Flags the current transform for re-computation.

Note: All set_* methods will automatically flag the component.

pub fn is_dirty(&self) -> bool

Returns whether or not the current transform is flagged for re-computation or "dirty".

pub fn matrix(&self) -> [[f32; 4]; 4]

Returns the local object matrix for the transform.

Combined with the parent's global Transform component it gives the global (or world) matrix for the current entity.