Struct amethyst::engine::state::StateMachine

pub struct StateMachine {
    // some fields omitted
}

A simple stack-based state machine.

Methods

impl StateMachine

fn new<T: 'static>(initial_state: T) -> StateMachine where T: State

fn current(&mut self) -> Option<&mut Box<State>>

Retrieves the currently active state.

fn start(&mut self)

Initializes the state machine.

fn handle_events(&mut self, events: &Vec<i32>)

Passes a vector of events to the active state to handle.

fn fixed_update(&mut self, delta_time: Duration)

Updates the currently active state at a steady, fixed interval.

fn update(&mut self, delta_time: Duration)

Updates the currently active state immediately.

fn switch<T: 'static>(&mut self, state: T) where T: State

Sets the currently active state.

fn push<T: 'static>(&mut self, state: T) where T: State

Pauses the active state (if any) and pushes a new state onto the state stack.

fn pop(&mut self)

Stops and removes the active state and un-pauses the next state on the stack (if any).

fn stop(&mut self)

Shuts the state machine down.