Trait amethyst::engine::state::State

pub trait State {
    fn new() -> Self where Self: Sized;

    fn on_start(&mut self) { ... }
    fn on_stop(&mut self) { ... }
    fn on_pause(&mut self) { ... }
    fn on_resume(&mut self) { ... }
    fn handle_events(&mut self, _events: &Vec<i32>) { ... }
    fn fixed_update(&mut self, _delta_time: Duration) { ... }
    fn update(&mut self, _delta_time: Duration) { ... }
}

A trait which defines game states that can be used by the state machine.

Required Methods

fn new() -> Self where Self: Sized

Provided Methods

fn on_start(&mut self)

Executed the first time when the game state is reached.

fn on_stop(&mut self)

Executed when the application finally exits.

fn on_pause(&mut self)

Executed when the application switches away to a different game state.

fn on_resume(&mut self)

Executed when the application returns to this game state once again.

fn handle_events(&mut self, _events: &Vec<i32>)

Executed on every frame before updating, for use in reacting to events. TODO: Replace i32 with an actual Event type of some kind.

fn fixed_update(&mut self, _delta_time: Duration)

Executed repeatedly at stable, predictable intervals (1/60th of a second by default).

fn update(&mut self, _delta_time: Duration)

Executed on every frame immediately, as fast as the engine will allow.