Struct amethyst::engine::StateMachine

pub struct StateMachine {
    // some fields omitted
}

A simple stack-based state machine (pushdown automaton).

Methods

impl StateMachine

fn new<T>(initial_state: T, planner: Planner<Arc<Mutex<Context>>>) -> StateMachine where T: State + 'static

fn is_running(&self) -> bool

Checks whether the state machine is running.

fn run_processors(&mut self, ctx: Arc<Mutex<Context>>)

Runs processors if the state machine is running.

fn start(&mut self, ctx: &mut Context)

Initializes the state machine.

Panics

Panics if no states are present in the stack.

fn handle_events(&mut self, events: &[Entity], ctx: &mut Context)

Passes a vector of events to the active state to handle.

fn fixed_update(&mut self, ctx: &mut Context)

Updates the currently active state at a steady, fixed interval.

fn update(&mut self, ctx: &mut Context)

Updates the currently active state immediately.