Module amethyst::gfx_device

Structs and enums holding graphics resources like gfx::Device, gfx::Factory, glutin::Window, etc.)

Modules

gfx_types
Graphics API resources that select their implementations at compile time.

Structs

DisplayConfig
Graphical display configuration.

GfxDevice
Holds all graphics resources required to render a Scene/Pipeline pair, except MainTarget.

MainTarget
Main render target that gets drawn on the screen.

Functions

video_init
Create a (GfxDevice, Factory, MainTarget) tuple from DisplayConfig