Macro amethyst::config

macro_rules! config {
    (
        // public configuration structure
        $(#[$identifier_meta:meta])*
        pub struct $identifier:ident {
            $(
                $(#[$field_meta:meta])*
                pub $field:ident: $ty:ty = $default:expr $(,)*
            )*
        }
    ) => { ... };
    (
        // private configuration structure
        $(#[$identifier_meta:meta])*
        struct $identifier:ident {
            $(
                $(#[$field_meta:meta])*
                pub $field:ident: $ty:ty = $default:expr $(,)*
            )*
        }
    ) => { ... };
    (
        // Implements the configuration structure's methods.
        impl $identifier:ident {
            $(
                $field:ident: $ty:ty = $default:expr,
            )*
        }
    ) => { ... };
}

The config! macro allows defining configuration structures that can load in .yml/.yaml files.

It implements the serde::Serialize, serde::Deserialize, and Config traits. As well as the standard libraries std::default::Default trait for the defaulting fields.

It follows Rust's syntax for defining structures with one addition, defaulting values. If the file does not contain a field of the configuration, then it will print out an error describing the issue and then load in this defaulting value.

In the case that the file is not found, or the configuration does not exist at all inside the file, then the configuration will have all default values.

Note: Documentation must be put on the line before the field, rather than on the same line.

Example Usage:

#[macro_use]
 extern crate amethyst;
 extern crate serde;
 #[macro_use]
 extern crate serde_derive;
 extern crate serde_yaml;
 extern crate toml;

config!(
    pub struct DisplayConfig {
        pub title: String = "Amethyst Game".to_string(),

        /// Resolution of the window size at start.
        pub resolution: (u32, u32) = (1920, 1080),
        
        pub fullscreen: bool = false,
    }
);