Enum amethyst::prelude::StateEvent

pub enum StateEvent<T = String> where
    T: Clone + Send + Sync + 'static,  {
    Window(Event),
    Ui(UiEvent),
    Input(InputEvent<T>),
}

The enum holding the different types of event that can be received in a State in the handle_event method.

Variants
Window(Event)

Events sent by the winit window.

Ui(UiEvent)

Events sent by the ui system.

Input(InputEvent<T>)

Events sent by the input system.