Enum amethyst::prelude::Trans

pub enum Trans<T, E> {
    None,
    Pop,
    Push(Box<dyn State<T, E>>),
    Switch(Box<dyn State<T, E>>),
    Quit,
}

Types of state transitions. T is the type of shared data between states. E is the type of events

Variants
None

Continue as normal.

Pop

Remove the active state and resume the next state on the stack or stop if there are none.

Push(Box<dyn State<T, E>>)

Pause the active state and push a new state onto the stack.

Switch(Box<dyn State<T, E>>)

Remove the current state on the stack and insert a different one.

Quit

Stop and remove all states and shut down the engine.