Module amethyst::processors::rendering

Default rendering processor types.

Structs

Camera
A Camera component. If this Camera is active then all changes in this component's fields will be applied to the camera that is being used to render the scene.

Light
A Light component. All changes in the light field will be applied to the associated renderer::Light in the frame.

Renderable
Entities with this component are rendered by the RenderingProcessor, modifying the transform field would affect the transform of the Fragment that is being rendered.

RendererConfig
A config required to create a rendering processor.

RenderingProcessor
A rendering processor struct.

Enums

Projection
A projection enum which is required to create a Camera component.