Struct amethyst::processors::rendering::Renderable

pub struct Renderable {
    pub translation: [f32; 3],
    pub rotation_axis: [f32; 3],
    pub rotation_angle: f32,
    pub scale: [f32; 3],
    // some fields omitted
}

Entities with this component are rendered by the RenderingProcessor, modifying the transform field would affect the transform of the Fragment that is being rendered.

Fields
translation: [f32; 3] rotation_axis: [f32; 3] rotation_angle: f32 scale: [f32; 3]
Methods

impl Renderable

fn new(mesh: &str, ka: &str, kd: &str) -> Renderable

Create a new Renderable component from names of assets loaded by context.asset_manager.