Struct amethyst::processors::transform::LocalTransform

pub struct LocalTransform { /* fields omitted */ }

Local position, rotation, and scale (from parent if it exists).

Methods

impl LocalTransform

fn translation(&self) -> [f32; 3]

fn rotation(&self) -> [f32; 4]

fn scale(&self) -> [f32; 3]

fn set_translation(&mut self, translation: [f32; 3])

fn set_rotation(&mut self, rotation: [f32; 4])

fn set_scale(&mut self, scale: [f32; 3])

fn set_translation_index(&mut self, index: usize, val: f32)

Set a specific part of the translation/position without modifying the others (must be an index less than 3).

Format: [0 = x, 1 = y, 2 = z]

e.g. transform.set_translation_index(1, 5.0) sets y to 5.0

fn set_rotation_index(&mut self, index: usize, val: f32)

Set a specific part of the rotation quaternion without modifying the others (must be an index less than 4).

Format: [0 = w, 1 = x, 2 = y, 3 = z]

e.g. transform.set_rot_index(1, 0.0) sets x to 0.0

fn set_scale_index(&mut self, index: usize, val: f32)

Set a specific part of the scale without modifying the others (must be an index less than 3).

Format: [0 = x, 1 = y, 2 = z]

e.g. transform.set_scale_index(2, 3.0) sets z to 3.0

fn flag(&self, dirty: bool)

Flags the current transform for re-computation.

Note: All set_* methods will automatically flag the component.

fn is_dirty(&self) -> bool

Returns whether or not the current transform is flagged for re-computation or "dirty".

fn matrix(&self) -> [[f32; 4]; 4]

Returns the local object matrix for the transform.

Combined with the parent's global Transform component it gives the global (or world) matrix for the current entity.