Module amethyst::renderer

High level rendering engine with multiple backends.

Reexports

pub use self::backend::Backend;

pub use self::frontend::{Renderable};

pub use self::types::{Buffer, BufferInfo, ClearMask, Shader};

Modules

backend
Makes low-level graphics API calls and manages memory.

frontend
Builds IR command buffers from Frames and feeds them into the backend.

ir
Platform-agnostic intermediate representation used by the frontend and backend to communicate.

types
Special types for representing cross-API GPU data.

Structs

Frame
A collection of renderable elements to be drawn by the Frontend.

Frontend
Simple renderer frontend.