Module amethyst::renderer::ir

Platform-agnostic intermediate representation used by the frontend and backend to communicate.

Structs

CommandEncoder
Builds a CommandBuffer and generates an associated SortKey.

CommandQueue
Queues and sorts CommandBuffers to minimize redundant state changes.

Enums

Command
Serialized graphics command.

Traits

AddCommands
A trait which adds methods to more easily populate a CommandBuffer.

Type Definitions

CommandBuffer
A collection of Commands.

SortKey
64-bit key used for sorting CommandBuffers. TODO: Need design for fields.