Trait amethyst::shred::System

pub trait System<'a> {
    type SystemData: DynamicSystemData<'a>;
    fn run(&mut self, data: Self::SystemData);

    fn running_time(&self) -> RunningTime { ... }

    fn accessor(&'b self) -> AccessorCow<'a, 'b, Self> { ... }

    fn setup(&mut self, res: &mut Resources) { ... }
}

A System, executed with a set of required Resources.

Associated Types

type SystemData: DynamicSystemData<'a>

The resource bundle required to execute this system.

You will mostly use a tuple of system data (which also implements SystemData). You can also create such a resource bundle by simply deriving SystemData for a struct.

Every SystemData is also a DynamicSystemData.

Required Methods

fn run(&mut self, data: Self::SystemData)

Executes the system with the required system data.

Provided Methods

fn running_time(&self) -> RunningTime

Returns a hint how long the system needs for running. This is used to optimize the way they're executed (might allow more parallelization).

Defaults to RunningTime::Average.

fn accessor(&'b self) -> AccessorCow<'a, 'b, Self>

Return the accessor from the [SystemData].

fn setup(&mut self, res: &mut Resources)

Sets up the Resources using Self::SystemData::setup.