Struct amethyst::shrev::EventChannel

pub struct EventChannel<E> { /* fields omitted */ }

Event channel

Methods

impl<E> EventChannel<E> where
E: Event,

pub fn new() -> EventChannel<E>

Create a new EventChannel with a default size of 50

pub fn with_capacity(size: usize) -> EventChannel<E>

Create a new EventChannel with the given starting capacity.

pub fn register_reader(&mut self) -> ReaderId<E>

Register a reader.

To be able to read events, a reader id is required. This is because otherwise the channel wouldn't know where in the ringbuffer the reader has read to earlier. This information is stored in the reader id.

pub fn slice_write(&mut self, events: &[E]) where
E: Clone,

Deprecated
: please use iter_write instead

Write a slice of events into storage

pub fn iter_write<I>(&mut self, iter: I) where
I: IntoIterator<Item = E>,
<I as IntoIterator>::IntoIter: ExactSizeIterator,

Write an iterator of events into storage

pub fn drain_vec_write(&mut self, events: &mut Vec<E>)

Drain a vector of events into storage.

pub fn single_write(&mut self, event: E)

Write a single event into storage.

Important traits for StorageIterator<'a, T>
impl<'a, T> Iterator for StorageIterator<'a, T> type Item = &'a T;

pub fn read(&self, reader_id: &mut ReaderId<E>) -> StorageIterator<E>

Read any events that have been written to storage since the readers last read.