Struct amethyst::ApplicationBuilder

pub struct ApplicationBuilder<'a, 'b, T> where
    T: State + 'static,  { /* fields omitted */ }

Helper builder for Applications.

Methods

impl<'a, 'b, T> ApplicationBuilder<'a, 'b, T> where
T: State + 'static,

fn new(initial_state: T, cfg: DisplayConfig) -> ApplicationBuilder<'a, 'b, T>

Creates a new ApplicationBuilder with the given initial game state and display configuration.

fn register<C>(self) -> ApplicationBuilder<'a, 'b, T> where
C: Component,

Registers a given component type.

fn add_barrier(self) -> ApplicationBuilder<'a, 'b, T>

Inserts a barrier which assures that all systems added before the barrier are executed before the ones after this barrier. Does nothing if there were no systems added since the last call to add_barrier(). Thread-local systems are not affected by barriers; they're always executed at the end.

fn with<S>(
self,
sys: S,
name: &str,
dep: &[&str]
) -> ApplicationBuilder<'a, 'b, T> where
S: System<'c> + Send + 'a + 'b,

Adds a given system sys, assigns it the string identifier name, and marks it dependent on systems dep. Note: all dependencies should be added before you add depending system

fn with_thread_local<S>(self, sys: S) -> ApplicationBuilder<'a, 'b, T> where
S: System<'c> + 'a + 'b,

Adds a given thread-local system sys All thread-local systems are executed sequentially after all non-thread-local systems

fn done(self) -> Application<'a, 'b>

Builds the Application and returns the result.