Struct amethyst::StateMachine

pub struct StateMachine<'a> { /* fields omitted */ }

A simple stack-based state machine (pushdown automaton).

Methods

impl<'a> StateMachine<'a>

pub fn new<S: State + 'a>(initial_state: S) -> StateMachine<'a>

Creates a new state machine with the given initial state.

pub fn is_running(&self) -> bool

Checks whether the state machine is running.

pub fn start(&mut self, engine: &mut World)

Initializes the state machine.

Panics

Panics if no states are present in the stack.

pub fn handle_event(&mut self, engine: &mut World, event: Event)

Passes a single event to the active state to handle.

pub fn fixed_update(&mut self, engine: &mut World)

Updates the currently active state at a steady, fixed interval.

pub fn update(&mut self, engine: &mut World)

Updates the currently active state immediately.