Enum amethyst::timing::Stopwatch

pub enum Stopwatch {
    Waiting,
    Started(Duration, Instant),
    Ended(Duration),
}

A stopwatch which accurately measures elapsed time.

Variants
Waiting

Initial state with an elapsed time value of 0 seconds.

Started(Duration, Instant)

Stopwatch has started counting the elapsed time since this Instant and accumuluated time from previous start/stop cycles Duration.

Ended(Duration)

Stopwatch has been stopped and reports the elapsed time Duration.

Methods

impl Stopwatch

pub fn new() -> Stopwatch

Creates a new stopwatch.

pub fn elapsed(&self) -> Duration

Retrieves the elapsed time.

pub fn restart(&mut self)

Stops, resets, and starts the stopwatch again.

pub fn start(&mut self)

Starts, or resumes, measuring elapsed time. If the stopwatch has been started and stopped before, the new results are compounded onto the existing elapsed time value.

Note: Starting an already running stopwatch will do nothing.

pub fn stop(&mut self)

Stops measuring elapsed time.

Note: Stopping a stopwatch that isn't running will do nothing.

pub fn reset(&mut self)

Clears the current elapsed time value.