Trait amethyst::EmptyState

pub trait EmptyState {
    fn on_start(&mut self, _data: StateData<()>) { ... }

    fn on_stop(&mut self, _data: StateData<()>) { ... }

    fn on_pause(&mut self, _data: StateData<()>) { ... }

    fn on_resume(&mut self, _data: StateData<()>) { ... }

    fn handle_event(
        &mut self, 
        _data: StateData<()>, 
        event: StateEvent
    ) -> EmptyTrans { ... }

    fn fixed_update(&mut self, _data: StateData<()>) -> EmptyTrans { ... }

    fn update(&mut self, _data: StateData<()>) -> EmptyTrans { ... }

    fn shadow_fixed_update(&mut self, _data: StateData<()>) { ... }

    fn shadow_update(&mut self, _data: StateData<()>) { ... }
}

An empty State trait. It contains no StateData or custom StateEvent.

Provided methods

fn on_start(&mut self, _data: StateData<()>)

Executed when the game state begins.

fn on_stop(&mut self, _data: StateData<()>)

Executed when the game state exits.

fn on_pause(&mut self, _data: StateData<()>)

Executed when a different game state is pushed onto the stack.

fn on_resume(&mut self, _data: StateData<()>)

Executed when the application returns to this game state once again.

fn handle_event(
&mut self,
_data: StateData<()>,
event: StateEvent
) -> EmptyTrans

Executed on every frame before updating, for use in reacting to events.

fn fixed_update(&mut self, _data: StateData<()>) -> EmptyTrans

Executed repeatedly at stable, predictable intervals (1/60th of a second by default).

fn update(&mut self, _data: StateData<()>) -> EmptyTrans

Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit).

fn shadow_fixed_update(&mut self, _data: StateData<()>)

Executed repeatedly at stable, predictable intervals (1/60th of a second by default), even when this is not the active state, as long as this state is on the StateMachine's state-stack.

fn shadow_update(&mut self, _data: StateData<()>)

Executed on every frame immediately, as fast as the engine will allow (taking into account the frame rate limit), even when this is not the active state, as long as this state is on the StateMachine's state-stack.

Loading content...