Struct amethyst::ecs::resources::InputHandler

pub struct InputHandler { /* fields omitted */ }

This struct holds state information about input devices.

For example, if a key is pressed on the keyboard, this struct will record that the key is pressed until it is released again. Usage requires pumping input events through this struct, using the following code in the handle_events method of State.

fn handle_events(&mut self,
                 events: &[WindowEvent],
                 _: &mut World,
                 _: &mut AssetManager,
                 _: &mut Pipeline)
                 -> Trans {
    // ...
    let mut input_handler = world.write_resource::<InputHandler>();
    input_handler.update(events);
    // ...
}

Methods

impl InputHandler

fn new() -> InputHandler

Creates a new input handler.

fn update(&mut self, events: &[WindowEvent])

Updates the input handler with new engine events.

fn pressed_keys(&self) -> PressedKeys

Returns an iterator for all the pressed down keys.

fn key_down(&self, key: VirtualKeyCode) -> bool

Checks if the given key is being pressed.

fn keys_down(&self, keys: &[VirtualKeyCode]) -> bool

Checks if all the given keys are being pressed.

fn key_once(&mut self, key: VirtualKeyCode) -> bool

Checks if the given key is being pressed and held down.

If key hasn't been released since the last key_once() query, this function will return false.

fn keys_once(&mut self, keys: &[VirtualKeyCode]) -> bool

Checks if the all the given keys are being pressed and held down.

If the keys haven't been released since the last key_once() query, this function will return false.

fn pressed_mouse_buttons(&self) -> PressedMouseButtons

Returns an iterator for all the pressed down mouse buttons.

fn mouse_button_down(&self, button: MouseButton) -> bool

Checks if the given mouse button is being pressed.

fn mouse_buttons_down(&self, buttons: &[MouseButton]) -> bool

Checks if all the given mouse buttons are being pressed.

fn mouse_button_once(&mut self, button: MouseButton) -> bool

Checks if the given mouse button is being pressed and held down.

If button hasn't been released since the last mouse_button_once() query, this function will return false.

fn mouse_buttons_once(&mut self, buttons: &[MouseButton]) -> bool

Checks if the all the given mouse buttons are being pressed and held down.

If the buttons haven't been released since the last mouse_button_once() query, this function will return false.