Struct amethyst::prelude::ApplicationBuilder

pub struct ApplicationBuilder<'a, 'b, T: State + 'a> {
    pub events: EventsLoop,
    // some fields omitted
}

Helper builder for Applications.

Fields
events: EventsLoop

Allows to create RenderSystem

Methods

impl<'a, 'b, T: State + 'a> ApplicationBuilder<'a, 'b, T>

pub fn new(initial_state: T) -> Result<Self>

Creates a new ApplicationBuilder with the given initial game state and display configuration.

pub fn register<C: Component>(self) -> Self

Registers a given component type.

pub fn add_resource<R>(self, res: R) -> Self where
R: Resource,

Adds an ECS resource which can be accessed from systems.

pub fn add_barrier(self) -> Self

Inserts a barrier which assures that all systems added before the barrier are executed before the ones after this barrier.

Does nothing if there were no systems added since the last call to add_barrier(). Thread-local systems are not affected by barriers; they're always executed at the end.

pub fn with<S>(self, sys: S, name: &str, dep: &[&str]) -> Self where
S: System<'c> + Send + 'a + 'b,

Adds a given system sys, assigns it the string identifier name, and marks it dependent on systems dep. Note: all dependencies should be added before you add depending system If you want to register systems which can not be specified as dependencies, you can use "" as their name, which will not panic (using another name twice will).

pub fn with_thread_local<S>(self, sys: S) -> Self where
S: System<'c> + 'a + 'b,

Adds a given thread-local system sys to the Application.

All thread-local systems are executed sequentially after all non-thread-local systems.

pub fn with_renderer(
self,
pipe: PipelineBuilder,
config: Option<DisplayConfig>
) -> Result<Self>

Automatically registers components, adds resources and the rendering system.

pub fn add_store<I, S>(self, name: I, store: S) -> Self where
I: Into<String>,
S: Store + Send + Sync + 'static,

Add asset loader to resources

pub fn register_asset<A, F>(self, make_context: F) -> Self where
A: Component + Asset + Clone + Send + Sync + 'static,
F: FnOnce(&mut World) -> A::Context,

Register new context within the loader

pub fn build(self) -> Result<Application<'a, 'b>>

Builds the Application and returns the result.