Module amethyst::renderer

High level rendering engine with multiple backends.

Reexports

pub use self::backend::Backend;

Modules

backend
Makes low-level graphics API calls and manages memory.

ir
Platform-agnostic intermediate representation used by the frontend and backend to communicate.

types
Special types for representing cross-API GPU data.

Structs

Frame
A collection of lights and objects to be rendered by the frontend.

Frontend
Simple renderer frontend.

Enums

Light
A light source.

Object
A physical renderable object.